Refs #10806: move implementation to cpp file
