+++ src/cpp/CMakeLists.txt
@@ -217,6 +217,7 @@ set(${PROJECT_NAME}_source_files
     utils/StringMatching.cpp
     utils/IPLocator.cpp
     utils/System.cpp
+    utils/SystemInfo.cpp
     utils/TimedConditionVariable.cpp
     utils/string_convert.cpp
