+++ test/unittest/rtps/common/CMakeLists.txt
@@ -16,10 +16,6 @@ if(NOT ((MSVC OR MSVC_IDE) AND EPROSIMA_INSTALLER))
     include(${PROJECT_SOURCE_DIR}/cmake/common/gtest.cmake)
     check_gtest()
 	
-	if(WIN32)
-		add_definitions(-DNOMINMAX)
-	endif(WIN32)
-
     if(GTEST_FOUND)
         set(CACHECHANGETESTS_SOURCE CacheChangeTests.cpp
             ${PROJECT_SOURCE_DIR}/src/cpp/rtps/common/Time_t.cpp)
