+++ test/unittest/statistics/dds/CMakeLists.txt
@@ -249,6 +249,7 @@ if(NOT ((MSVC OR MSVC_IDE) AND EPROSIMA_INSTALLER))
                 ${PROJECT_SOURCE_DIR}/src/cpp/utils/md5.cpp
                 ${PROJECT_SOURCE_DIR}/src/cpp/utils/StringMatching.cpp
                 ${PROJECT_SOURCE_DIR}/src/cpp/utils/string_convert.cpp
+                ${PROJECT_SOURCE_DIR}/src/cpp/utils/SystemInfo.cpp
                 ${PROJECT_SOURCE_DIR}/src/cpp/utils/TimedConditionVariable.cpp
                 )
